requirements.txt
discord.py==2.4.0
python-dotenv==1.0.1
Requests==2.32.3
slack_bolt==1.22.0
